# Conflicts:
#	Katon_Collection/Assets/Anima2D.meta
#	Katon_Collection/Assets/Scripts/ServerInterface/SI_Player.cs
#	Katon_Collection/Library/CurrentLayout-default.dwlt
#	Katon_Collection/Library/CurrentMaximizeLayout.dwlt
#	Katon_Collection/Library/ScriptAssemblies/Assembly-CSharp.dll
#	Katon_Collection/Library/StateCache/SceneView/f2e1c3-mainStage.json
#	Katon_Collection/Library/assetDatabase3
#	Katon_Collection/Library/expandedItems
#	Katon_Collection/Library/metadata/15/15067e59d2bb1fe4da5453c0fa02c3a1.info
#	Katon_Collection/Library/metadata/30/30d430197e5a35c4185fea5ec0e34623
#	Katon_Collection/Library/metadata/48/48a57341e3a1f3943b2277029b373d51.info
#	Katon_Collection/Library/metadata/4b/4b5865ddddd8d5548b66d88ecdd2489a
#	Katon_Collection/Library/metadata/60/60f98cabfa3ce2a4d981cf40a7b0a9d6
#	Katon_Collection/Library/metadata/70/701984e2c0a630c44a068946cec4b3ee
#	Katon_Collection/Library/metadata/78/78e932ee174007c44b368b696d443d0e
#	Katon_Collection/Library/metadata/7c/7c2738285fc20bc44819f15ca6e8ea3e
#	Katon_Collection/Library/metadata/7e/7e43d0e6254d964468c424dd10fa22dc
#	Katon_Collection/Library/metadata/92/9298fc38c80836b4c88f519050fb2597
#	Katon_Collection/Library/metadata/a6/a6ca78af4c90b0a875dad64c384e38ab
#	Katon_Collection/Library/metadata/cf/cf90a4b194b0a8049b871366df6e3cc6.info
#	Katon_Collection/Library/metadata/d0/d0e1db54a370655478bfe5dd9b15b234
#	Katon_Collection/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache

+++ Katon_Collection/Assets/Scripts/ServerInterface/SI_Player.cs
@@ -14,7 +14,7 @@ public class SI_Player : Photon.MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        for (int i = 0; i < (int)placePoint.Length; i++)
+        for (int i = 0; i < placePoint.Length; i++)
         {
             placePoint[i] = 0;
         }
